[GRAPH-GET]: Allow node as starting node and some refactor
CHANGES:
 - Adjust unit test to cover some changes
